Rename area -> spawn region
