--- Content.Client/_ES/Area/ESAreaSystem.cs
@@ -1,5 +1,0 @@
-using Content.Shared._ES.Area;
-
-namespace Content.Client._ES.Area;
-
-public sealed class ESAreaSystem : ESSharedAreaSystem;
