+++ Content.Client/_ES/SpawnRegion/ESSpawnRegionSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared._ES.SpawnRegion;
+
+namespace Content.Client._ES.SpawnRegion;
+
+public sealed class ESSpawnRegionSystem : ESSharedSpawnRegionSystem;
