+++ Resources/Maps/_ES/estestship_grid.yml
@@ -2184,7 +2184,7 @@ entities:
     - type: Transform
       pos: -3.5,-23.5
       parent: 1
-- proto: ESMarkerAreaMaintenance
+- proto: ESMarkerSpawnRegionMaintenance
   entities:
   - uid: 761
     components:
