--- Resources/Prototypes/_ES/areas.yml
@@ -1,2 +1,0 @@
-- type: esArea
-  id: ESMaintenance
