+++ Resources/Prototypes/_ES/spawn_regions.yml
@@ -1,0 +1,2 @@
+- type: esSpawnRegion
+  id: ESMaintenance
